pyproject.toml
[tool.poetry]
name = "chess-engine"
version = "0.1.0"
description = "A chess engine using neural networks trained through self-play"
authors = ["Leonardo Mannini <leonardo.mannini@studio.unibo.it>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
numpy = "^1.26.0"
python-chess = "^1.999"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-multipart = "^0.0.9"
pydantic = "^2.6.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100 